requirements.txt
Flask==3.0.3
Flask-WTF==1.2.2
WTForms==3.1.2
email-validator==2.2.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
